Fix problems with the use of 32-bit integers for offsets. (See #9514)
This commit fixes both the issue that Alex Herbert raised on ome-devel
and an issue with OMERO.insight or OmeroJava being able to download
files > 2.1GB.

+++ components/blitz/src/ome/formats/OMEROMetadataStoreClient.java
@@ -1863,7 +1863,7 @@ public class OMEROMetadataStoreClient
                 stream = new FileInputStream(file);
                 rawFileStore.setFileId(originalFile.getId().getValue());
                 int rlen = 0;
-                int offset = 0;
+                long offset = 0;
                 while (stream.available() != 0)
                 {
                     rlen = stream.read(buf);
